+++ src/scitex_clew/_viz/_json.py
@@ -218,7 +218,7 @@ def _generate_dag_json_multi(
 
         dag = verify_claims_dag()
     elif target_files:
-        from .._dag import verify_dag
+        from .._chain import verify_dag
 
         dag = verify_dag(target_files)
     else: